Melhoria em Contas

+++ src/main/java/br/com/marcielli/BancoM/controller/UserClienteController.java
@@ -1,12 +1,10 @@
 package br.com.marcielli.BancoM.controller;
 
 import java.util.List;
-import java.util.Set;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,21 +12,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ResponseStatusException;
-
-import br.com.marcielli.BancoM.dto.ClienteMapper;
-import br.com.marcielli.BancoM.dto.ClienteResponseDTO;
 import br.com.marcielli.BancoM.dto.security.UserClienteResponseDTO;
 import br.com.marcielli.BancoM.dto.security.UserCreateDTO;
 import br.com.marcielli.BancoM.entity.Cliente;
 import br.com.marcielli.BancoM.entity.Endereco;
-import br.com.marcielli.BancoM.entity.Role;
 import br.com.marcielli.BancoM.entity.User;
-import br.com.marcielli.BancoM.repository.ClienteRepository;
-import br.com.marcielli.BancoM.repository.RoleRepository;
 import br.com.marcielli.BancoM.repository.UserRepository;
 import br.com.marcielli.BancoM.service.UserClienteService;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 
 @RestController
@@ -36,7 +26,7 @@ public class UserClienteController {
 
 	private final UserRepository userRepository;
 	private final UserClienteService clienteService;
-	
+
 	public UserClienteController(UserClienteService clienteService, UserRepository userRepository) {
 		this.clienteService = clienteService;
 		this.userRepository = userRepository;
@@ -44,111 +34,114 @@ public class UserClienteController {
 
 	@PostMapping("/users")
 	@Transactional
-	public ResponseEntity<String> newUser(@RequestBody UserCreateDTO dto){
-			
+	public ResponseEntity<String> newUser(@RequestBody UserCreateDTO dto) {
+
 		User clienteAdicionado = clienteService.save(dto);
-		
-		if(clienteAdicionado != null) {
+
+		if (clienteAdicionado != null) {
 			return new ResponseEntity<String>("Cliente adicionado com sucesso", HttpStatus.CREATED);
 		} else {
 			return new ResponseEntity<String>("Tente novamente mais tarde.", HttpStatus.NOT_ACCEPTABLE);
 		}
 	}
-	
+
 	@GetMapping("/users")
 	@PreAuthorize("hasAuthority('SCOPE_ADMIN')")
-	public ResponseEntity<List<User>> listUsers() {		
+	public ResponseEntity<List<User>> listUsers() {
 		var users = userRepository.findAll();
 		return ResponseEntity.ok(users);
 	}
-	
+
 	@GetMapping("/users/{id}")
 	@PreAuthorize("hasAuthority('SCOPE_ADMIN') or hasAuthority('SCOPE_BASIC')")
 	@Transactional
 	public ResponseEntity<?> getClienteById(@PathVariable("id") Long id) {
-		
+
 		Cliente clienteUnico = clienteService.getClienteById(id);
-		
+
 		if (clienteUnico == null || clienteUnico.getUser() == null) {
-	        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
-	    }
-		
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
+		}
+
 		boolean isAdmin = clienteUnico.getUser().getRoles().stream()
-			    .anyMatch(role -> "ADMIN".equalsIgnoreCase(role.getName()));
-		
-		if(!isAdmin) {
+				.anyMatch(role -> "ADMIN".equalsIgnoreCase(role.getName()));
+
+		if (!isAdmin) {
 			UserClienteResponseDTO response = new UserClienteResponseDTO();
 			response.setId(id);
 			response.setNome(clienteUnico.getNome());
 			response.setCpf(clienteUnico.getCpf());
-			
+
 			Endereco endereco = clienteUnico.getEndereco();
-	        if (endereco != null) {
-	            response.setCep(endereco.getCep());
-	            response.setCidade(endereco.getCidade());
-	            response.setEstado(endereco.getEstado());
-	            response.setRua(endereco.getRua());
-	            response.setNumero(endereco.getNumero());
-	            response.setBairro(endereco.getBairro());
-	            response.setComplemento(endereco.getComplemento());
-	        }
-			
+			if (endereco != null) {
+				response.setCep(endereco.getCep());
+				response.setCidade(endereco.getCidade());
+				response.setEstado(endereco.getEstado());
+				response.setRua(endereco.getRua());
+				response.setNumero(endereco.getNumero());
+				response.setBairro(endereco.getBairro());
+				response.setComplemento(endereco.getComplemento());
+			}
+
 			return ResponseEntity.status(HttpStatus.OK).body(response);
-		} else {			
+		} else {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
 		}
 	}
-	
-	@PutMapping("/users/{id}") 
+
+	@PutMapping("/users/{id}")
 	@PreAuthorize("hasAuthority('SCOPE_ADMIN') or hasAuthority('SCOPE_BASIC')")
 	@Transactional
 	public ResponseEntity<?> atualizar(@PathVariable("id") Long id, @RequestBody UserCreateDTO dto) {
 
 		Cliente clienteUnico = clienteService.update(id, dto);
-		
+
 		if (clienteUnico == null || clienteUnico.getUser() == null) {
-	        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
-	    }
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
+		}
 		boolean isAdmin = clienteUnico.getUser().getRoles().stream()
-			    .anyMatch(role -> "ADMIN".equalsIgnoreCase(role.getName()));
-		
-		if(!isAdmin) {
+				.anyMatch(role -> "ADMIN".equalsIgnoreCase(role.getName()));
+
+		if (!isAdmin) {
 			UserClienteResponseDTO response = new UserClienteResponseDTO();
 			response.setId(id);
 			response.setNome(clienteUnico.getNome());
 			response.setCpf(clienteUnico.getCpf());
-			
+
 			Endereco endereco = clienteUnico.getEndereco();
-	        if (endereco != null) {
-	            response.setCep(endereco.getCep());
-	            response.setCidade(endereco.getCidade());
-	            response.setEstado(endereco.getEstado());
-	            response.setRua(endereco.getRua());
-	            response.setNumero(endereco.getNumero());
-	            response.setBairro(endereco.getBairro());
-	            response.setComplemento(endereco.getComplemento());
-	        }
-			
+			if (endereco != null) {
+				response.setCep(endereco.getCep());
+				response.setCidade(endereco.getCidade());
+				response.setEstado(endereco.getEstado());
+				response.setRua(endereco.getRua());
+				response.setNumero(endereco.getNumero());
+				response.setBairro(endereco.getBairro());
+				response.setComplemento(endereco.getComplemento());
+			}
+
 			return ResponseEntity.status(HttpStatus.OK).body(response);
-		} else {			
+		} else {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
 		}
-		
+
 	}
-	
+
 	@DeleteMapping("/users/{id}")
 	@PreAuthorize("hasAuthority('SCOPE_ADMIN') or hasAuthority('SCOPE_BASIC')")
 	@Transactional
 	public ResponseEntity<?> deletar(@PathVariable("id") Long id) {
+		
+		if(id == null) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O cliente não existe!");
+		}
 
-	    boolean clienteUnico = clienteService.delete(id);
+		boolean clienteUnico = clienteService.delete(id);
 
-	    if (clienteUnico) {
-	        return ResponseEntity.status(HttpStatus.OK).body("Cliente deletado com sucesso!");
-	    } else {
-	        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Erro. Tente novamente mais tarde.");
-	    }
+		if (clienteUnico) {
+			return ResponseEntity.status(HttpStatus.OK).body("Cliente deletado com sucesso!");
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Erro. Tente novamente mais tarde.");
+		}
 	}
 
-	
 }
